+++ src/api/main.ts
@@ -4,6 +4,7 @@
  */
 
 import { Hono } from "hono";
+import { registerContactRoutes } from "@/api/contact/contact.routes.ts";
 import { registerHealthRoutes } from "@/api/health/health.routes.ts";
 import { registerProductRoutes } from "@/api/products/product.routes.ts";
 import { registerUserRoutes } from "@/api/users/user.routes.ts";
@@ -13,5 +14,6 @@ const api = new Hono();
 registerHealthRoutes(api);
 registerUserRoutes(api);
 registerProductRoutes(api);
+registerContactRoutes(api);
 
 export default api;